Initial commit: vapor kind of works
This commit adds a working example which will bootstrap an execution
environment in Docker or Kubernetes, spin up workers, and then executes
a function in parallel on the workers. After the initial bootstrapping,
the workers stay running and wait for future tasks. Everything is still
really buggy.

This commit also adds a README and a LICENSE.